requirements.txt
cs50
Flask
Flask-Session
requests
IMDbPY
scikit-learn